+++ app/controllers/report_activities_controller.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+class ReportActivitiesController < BaseController
+  include Secured
+  include Reports::Breadcrumbed
+
+  def show
+    @report = Report.find(params["report_id"])
+    authorize @report
+
+    prepare_default_report_trail @report
+
+    @report_presenter = ReportPresenter.new(@report)
+    @updated_activities = @report.activities_updated
+    @new_activities = @report.new_activities
+
+    render "reports/activities"
+  end
+end